Refactor MarketSnapshot and JSON serialization
- Simplified the `parse_rpc_any` function in `jsonrpc.rs` to use a more concise mapping for request parsing.
- Updated `MarketSnapshot` in `models.rs` to replace individual indicator fields with a unified `indicators` HashMap, allowing for a more flexible and extensible design.
- Introduced legacy-compatible accessors for indicators to maintain backward compatibility during the transition to the new structure.
- Removed deprecated fields from `MarketSnapshot` and adjusted tests accordingly to validate the new structure.
- Added property tests for normalized indicators to ensure correct mapping and state labels.
- Improved readability and maintainability of tests by using more idiomatic Rust patterns.

+++ crates/shared/src/normalized/candle_generator.rs
@@ -234,7 +234,7 @@ mod tests {
                 trade_id: format!("t_{}", i),
             };
             let (_closed, live) = generator.process_trade(&trade);
-            assert_eq!(live.trades_count, (i + 1) as u64);
+            assert_eq!(live.trades_count, i + 1);
         }
     }
 }
